feat: 드롭 시 전체 렌더링 최적화 중, 폴더 추가 및 정리중.
