+++ .github/pull_request_template.md
@@ -4,10 +4,10 @@
 
 - [ ] 배포 후 url 제출
 
-- [ ] API 호출 최적화(`Promise.all` 이해)
+- [x] API 호출 최적화(`Promise.all` 이해)
 
-- [ ] SearchDialog 불필요한 연산 최적화
-- [ ] SearchDialog 불필요한 리렌더링 최적화
+- [x] SearchDialog 불필요한 연산 최적화
+- [x] SearchDialog 불필요한 리렌더링 최적화
 
 - [ ] 시간표 블록 드래그시 렌더링 최적화
 - [ ] 시간표 블록 드롭시 렌더링 최적화
@@ -17,6 +17,8 @@
 <!-- 과제에 대한 회고를 작성해주세요 -->
 
 ### 기술적 성장
+
+
 <!-- 예시
 - 새로 학습한 개념
 - 기존 지식의 재발견/심화
