+++ src/ScheduleDndProvider.tsx
@@ -7,7 +7,7 @@ import {
   useSensors,
 } from '@dnd-kit/core';
 import { PropsWithChildren, useCallback } from 'react';
-import { CellSize, DAY_LABELS } from './constants.ts';
+import { CellSize, DAY_LABELS } from './constants/constants.ts';
 import { Schedule } from './types.ts';
 import { useDragState } from './SchedulesDragStateProvider.tsx';
 
@@ -53,11 +53,11 @@ interface Props extends PropsWithChildren {
   onScheduleUpdate: (tableId: string, schedules: Schedule[]) => void;
 }
 
-export default function ScheduleDndProvider({ 
-  children, 
-  tableId, 
+export default function ScheduleDndProvider({
+  children,
+  tableId,
   schedules,
-  onScheduleUpdate 
+  onScheduleUpdate,
 }: Props) {
   const { setActiveTableId, setIsDragging } = useDragState();
   const sensors = useSensors(
@@ -68,49 +68,49 @@ export default function ScheduleDndProvider({
     })
   );
 
-  const handleDragStart = useCallback(
-    (event: DragStartEvent) => {
-      const [dragTableId] = String(event.active.id).split(':');
-      setActiveTableId(dragTableId);
-      setIsDragging(true);
-    },
-    []
-  );
+  const handleDragStart = useCallback((event: DragStartEvent) => {
+    const [dragTableId] = String(event.active.id).split(':');
+    setActiveTableId(dragTableId);
+    setIsDragging(true);
+  }, []);
 
-  const handleDragEnd = useCallback((event: any) => {
-    const { active, delta } = event;
-    const { x, y } = delta;
-    const [dragTableId, index] = active.id.split(':');
-    
-    // 이 테이블의 드래그인 경우만 처리
-    if (dragTableId !== tableId) {
+  const handleDragEnd = useCallback(
+    (event: any) => {
+      const { active, delta } = event;
+      const { x, y } = delta;
+      const [dragTableId, index] = active.id.split(':');
+
+      // 이 테이블의 드래그인 경우만 처리
+      if (dragTableId !== tableId) {
+        setActiveTableId(null);
+        setIsDragging(false);
+        return;
+      }
+
+      const schedule = schedules[Number(index)];
+      const nowDayIndex = DAY_LABELS.indexOf(
+        schedule.day as (typeof DAY_LABELS)[number]
+      );
+      const moveDayIndex = Math.floor(x / 80);
+      const moveTimeIndex = Math.floor(y / 30);
+
+      const updatedSchedules = schedules.map((targetSchedule, targetIndex) => {
+        if (targetIndex !== Number(index)) {
+          return targetSchedule;
+        }
+        return {
+          ...targetSchedule,
+          day: DAY_LABELS[nowDayIndex + moveDayIndex],
+          range: targetSchedule.range.map(time => time + moveTimeIndex),
+        };
+      });
+
+      onScheduleUpdate(tableId, updatedSchedules);
       setActiveTableId(null);
       setIsDragging(false);
-      return;
-    }
-
-    const schedule = schedules[Number(index)];
-    const nowDayIndex = DAY_LABELS.indexOf(
-      schedule.day as (typeof DAY_LABELS)[number]
-    );
-    const moveDayIndex = Math.floor(x / 80);
-    const moveTimeIndex = Math.floor(y / 30);
-
-    const updatedSchedules = schedules.map((targetSchedule, targetIndex) => {
-      if (targetIndex !== Number(index)) {
-        return targetSchedule;
-      }
-      return {
-        ...targetSchedule,
-        day: DAY_LABELS[nowDayIndex + moveDayIndex],
-        range: targetSchedule.range.map(time => time + moveTimeIndex),
-      };
-    });
-
-    onScheduleUpdate(tableId, updatedSchedules);
-    setActiveTableId(null);
-    setIsDragging(false);
-  }, [tableId, schedules, onScheduleUpdate]);
+    },
+    [tableId, schedules, onScheduleUpdate]
+  );
 
   return (
     <DndContext
