+++ src/components/schedules/SchedulesTable/ScheduleTables.tsx
@@ -1,132 +1,13 @@
 import { Flex } from '@chakra-ui/react/flex';
-import { Stack } from '@chakra-ui/react/stack';
-import { Heading } from '@chakra-ui/react/typography';
-import { Button, ButtonGroup } from '@chakra-ui/react/button';
 import { useScheduleContext } from '../../../ScheduleContext.tsx';
-import React, { useState, useCallback, useMemo, lazy, memo } from 'react';
-
-import ScheduleDndProvider from '../../../ScheduleDndProvider.tsx';
-import { DragStateProvider } from '../../../SchedulesDragStateProvider.tsx';
+import { useState, useCallback, useMemo } from 'react';
 import { Schedule } from '../../../types.ts';
-const ScheduleTable = lazy(() => import('./ScheduleTable.tsx'));
-
-interface LazyComponentWithPreload
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  extends React.LazyExoticComponent<React.ComponentType<any>> {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  preload: () => Promise<any>;
-}
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function lazyWithPreloading(importFn: () => Promise<any>) {
-  const Component = React.lazy(importFn) as LazyComponentWithPreload;
-  Component.preload = importFn;
-  return Component;
-}
+import { ScheduleTableContainer } from './components';
+import { lazyWithPreloading } from '../../../lib/lazyWithPreloading.ts';
 const SearchDialog = lazyWithPreloading(
   () => import('../SearchDialog/SearchDialog.tsx')
 );
 
-const TableButtonGroup = memo(
-  ({
-    tableId,
-    disabled,
-    onSearch,
-    onDuplicate,
-    onRemove,
-  }: {
-    tableId: string;
-    disabled: boolean;
-    onSearch: (tableId: string) => void;
-    onDuplicate: (tableId: string) => void;
-    onRemove: (tableId: string) => void;
-  }) => {
-    const handleSearch = useCallback(
-      () => onSearch(tableId),
-      [tableId, onSearch]
-    );
-    const handleDuplicate = useCallback(
-      () => onDuplicate(tableId),
-      [tableId, onDuplicate]
-    );
-    const handleRemove = useCallback(
-      () => onRemove(tableId),
-      [tableId, onRemove]
-    );
-    return (
-      <ButtonGroup size="sm" isAttached>
-        <Button colorScheme="green" onClick={handleSearch}>
-          시간표 추가
-        </Button>
-        <Button colorScheme="green" mx="1px" onClick={handleDuplicate}>
-          복제
-        </Button>
-        <Button
-          colorScheme="green"
-          isDisabled={disabled}
-          onClick={handleRemove}
-        >
-          삭제
-        </Button>
-      </ButtonGroup>
-    );
-  }
-);
-const ScheduleTableItem = memo(
-  ({
-    tableId,
-    schedules,
-    index,
-    disabledRemoveButton,
-    onSearchOpen,
-    onDuplicate,
-    onRemove,
-    onScheduleTimeClick,
-    onDeleteButtonClick,
-    onScheduleUpdate,
-  }: {
-    tableId: string;
-    schedules: Schedule[];
-    index: number;
-    disabledRemoveButton: boolean;
-    onSearchOpen: (tableId: string) => void;
-    onDuplicate: (tableId: string) => void;
-    onRemove: (tableId: string) => void;
-    onScheduleTimeClick: (timeInfo: { day?: string; time?: number }) => void; // 변경!
-    onDeleteButtonClick: ({ day, time }: { day: string; time: number }) => void; // 변경!
-    onScheduleUpdate: (tableId: string, schedules: Schedule[]) => void;
-  }) => {
-    return (
-      <DragStateProvider>
-        <ScheduleDndProvider
-          tableId={tableId}
-          schedules={schedules}
-          onScheduleUpdate={onScheduleUpdate}
-        >
-          <Stack width="600px">
-            <Flex justifyContent="space-between" alignItems="center">
-              <Heading as="h3" fontSize="lg">
-                시간표 {index + 1}
-              </Heading>
-              <TableButtonGroup
-                tableId={tableId}
-                disabled={disabledRemoveButton}
-                onSearch={onSearchOpen}
-                onDuplicate={onDuplicate}
-                onRemove={onRemove}
-              />
-            </Flex>
-            <ScheduleTable
-              schedules={schedules}
-              tableId={tableId}
-              onScheduleTimeClick={onScheduleTimeClick}
-              onDeleteButtonClick={onDeleteButtonClick}
-            />
-          </Stack>
-        </ScheduleDndProvider>
-      </DragStateProvider>
-    );
-  }
-);
 const ScheduleTables = () => {
   const { schedulesMap, setSchedulesMap } = useScheduleContext();
   const [searchInfo, setSearchInfo] = useState<{
@@ -173,10 +54,19 @@ const ScheduleTables = () => {
 
   const handleScheduleUpdate = useCallback(
     (tableId: string, updatedSchedules: Schedule[]) => {
-      setSchedulesMap(prev => ({
-        ...prev,
-        [tableId]: updatedSchedules,
-      }));
+      setSchedulesMap(
+        prev => {
+          const next: Record<string, Schedule[]> = {};
+          for (const key of Object.keys(prev)) {
+            next[key] = key === tableId ? updatedSchedules : prev[key];
+          }
+          return next;
+        }
+        //   ({
+        //   ...prev,
+        //   [tableId]: updatedSchedules,
+        // })
+      );
     },
     [setSchedulesMap]
   );
@@ -219,7 +109,6 @@ const ScheduleTables = () => {
         },
       };
     });
-
     return handlers;
   }, []); // 의존성 배열 비움!
 
@@ -227,7 +116,7 @@ const ScheduleTables = () => {
     <>
       <Flex w="full" gap={6} p={6} flexWrap="wrap">
         {scheduleEntries.map(([tableId, schedules], index) => (
-          <ScheduleTableItem
+          <ScheduleTableContainer
             key={tableId}
             tableId={tableId}
             schedules={schedules}
