+++ src/components/schedules/SchedulesTable/components/ScheduleTableHeader.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { GridItem } from '@chakra-ui/react/grid';
 import { Flex } from '@chakra-ui/react/flex';
 import { Text } from '@chakra-ui/react/typography';
-import { DAY_LABELS } from '../../../../../constants.ts';
+import { DAY_LABELS } from '../../../../constants/constants.ts';
 
 export const ScheduleTableHeader = memo(() => (
   <>
