+++ src/components/schedules/SchedulesTable/components/TimeRow.tsx
@@ -1,36 +1,38 @@
 import { Fragment, memo, useCallback } from 'react';
 import { GridItem } from '@chakra-ui/react/grid';
 import { Flex } from '@chakra-ui/react/flex';
-import { fill2 } from '../../../../utils.ts';
-import { DAY_LABELS } from '../../../../constants.ts';
+import { fill2 } from '../../../../utils/utils.ts';
+import { DAY_LABELS } from '../../../../constants/constants.ts';
 
 // 개별 셀 컴포넌트로 분리
-const TimeCell = memo(({
-  day,
-  timeNumber,
-  isEvening,
-  onTimeClick
-}: {
-  day: string;
-  timeNumber: number;
-  isEvening: boolean;
-  onTimeClick: (day: string, time: number) => void;
-}) => {
-  const handleClick = useCallback(() => {
-    onTimeClick(day, timeNumber);
-  }, [day, timeNumber, onTimeClick]);
+const TimeCell = memo(
+  ({
+    day,
+    timeNumber,
+    isEvening,
+    onTimeClick,
+  }: {
+    day: string;
+    timeNumber: number;
+    isEvening: boolean;
+    onTimeClick: (day: string, time: number) => void;
+  }) => {
+    const handleClick = useCallback(() => {
+      onTimeClick(day, timeNumber);
+    }, [day, timeNumber, onTimeClick]);
 
-  return (
-    <GridItem
-      borderWidth="1px 0 0 1px"
-      borderColor="gray.300"
-      bg={isEvening ? 'gray.100' : 'white'}
-      cursor="pointer"
-      _hover={{ bg: 'yellow.100' }}
-      onClick={handleClick}
-    />
-  );
-});
+    return (
+      <GridItem
+        borderWidth="1px 0 0 1px"
+        borderColor="gray.300"
+        bg={isEvening ? 'gray.100' : 'white'}
+        cursor="pointer"
+        _hover={{ bg: 'yellow.100' }}
+        onClick={handleClick}
+      />
+    );
+  }
+);
 
 export const TimeRow = memo(
   ({
