+++ src/components/schedules/SchedulesTable/components/index.ts
@@ -1,1 +1,5 @@
 export * from './TimeRow';
+export * from './ScheduleGrid';
+export * from './ScheduleTableHeader';
+export * from './ScheduleTableContainer';
+export * from './SchedulesTableButton';
