+++ src/components/schedules/SearchDialog/components/CheckDays.tsx
@@ -5,7 +5,7 @@ import {
   FormControl,
   FormLabel,
 } from '@chakra-ui/react';
-import { DAY_LABELS } from '../../../../constants.ts';
+import { DAY_LABELS } from '../../../../constants/constants.ts';
 import { SearchOption } from '../SearchDialog.tsx';
 import { memo } from 'react';
