+++ src/components/schedules/SearchDialog/components/CheckTime.tsx
@@ -4,7 +4,7 @@ import { Checkbox, CheckboxGroup } from '@chakra-ui/react/checkbox';
 import { Wrap } from '@chakra-ui/react/wrap';
 import { Tag, TagCloseButton, TagLabel } from '@chakra-ui/react/tag';
 import { Stack } from '@chakra-ui/react/stack';
-import { TIME_SLOTS } from '../../../../constants.ts';
+import { TIME_SLOTS } from '../../../../constants/constants.ts';
 import { Box } from '@chakra-ui/react/box';
 
 import { FormControl } from '@chakra-ui/react';
