+++ src/constants/constants.ts
@@ -1,4 +1,4 @@
-import { parseHnM } from './utils.ts';
+import { parseHnM } from '../utils/utils.ts';
 
 export const DAY_LABELS = ['월', '화', '수', '목', '금', '토'] as const;
